server/requirements.txt
livekit-agents
livekit-plugins-google
livekit-plugins-silero 
python-dotenv
 
livekit-api
uvicorn
fastapi
pydantic
google-generativeai